# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Kubernetes FileBrowser",
        "title": "Kube-FileBrowser Swagger",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/filebrowser/createdir": {
            "post": {
                "description": "容器文件浏览器 - 创建目录",
                "tags": [
                    "FileBrowser"
                ],
                "summary": "CreateDir",
                "parameters": [
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "old_path",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "path",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "old_path",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "path",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "old_path",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "path",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "old_path",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "path",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.JSONResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.JSONResult"
                        }
                    }
                }
            }
        },
        "/api/filebrowser/createfile": {
            "post": {
                "description": "容器文件浏览器 - 创建文件",
                "tags": [
                    "FileBrowser"
                ],
                "summary": "CreateFile",
                "parameters": [
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "old_path",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "path",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "old_path",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "path",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "old_path",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "path",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "old_path",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "path",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "内容",
                        "name": "content",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.JSONResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.JSONResult"
                        }
                    }
                }
            }
        },
        "/api/filebrowser/list": {
            "get": {
                "description": "容器文件浏览器 - 文件列表",
                "tags": [
                    "FileBrowser"
                ],
                "summary": "ListFile",
                "parameters": [
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "old_path",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "path",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "old_path",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "path",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "old_path",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "path",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "container",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "old_path",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "path",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "pod",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.JSONResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.JSONResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.JSONResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string",
                    "example": "消息"
                }
            }
        }
    }
}